fix type

+++ sqlbuilder.go
@@ -63,7 +63,7 @@ func escapeStr(s string) string {
 func QueriesBuild(
 	data [][]string,
 	querytemplate string,
-	maxallowedpack int) (queries []string, err error) {
+	maxallowedpack uint64) (queries []string, err error) {
 	if maxallowedpack < MINALLOWEDPACKETLEN {
 		err = errors.New("max_allowed_packet can't be less than 1024")
 		return nil, err
@@ -82,7 +82,7 @@ func QueriesBuild(
 	outsql := &strings.Builder{}
 	outsql.WriteString(SQLQuery)
 	outsql.WriteString(rowBuild(data[0]))
-	if outsql.Len() > maxallowedpack {
+	if uint64(outsql.Len()) > maxallowedpack {
 		err = fmt.Errorf("query is too big - max_allowed_packet limit is %d", maxallowedpack)
 		return nil, err
 	}
@@ -93,7 +93,7 @@ func QueriesBuild(
 	}
 	for i := 1; i < len(data); i++ {
 		r := rowBuild(data[i])
-		if (outsql.Len() + len(r) + 1) >= maxallowedpack {
+		if uint64(outsql.Len()+len(r)+1) >= maxallowedpack {
 			queries = append(queries, outsql.String())
 			outsql.Reset()
 			outsql.WriteString(SQLQuery)
